@@ -70,96 +70,96 @@
     "https": "1.0.0",
     "jwks-rsa": "^1.6.0",
     "minimist": "^1.2.0",
-    "mongodb": "^3.2.7",
-    "mongoose": "^5.6.3",
+    "mongodb": "^3.3.2",
+    "mongoose": "^5.7.1",
     "mongoose-plugin-autoinc": "^1.1.9",
     "morgan": "^1.9.1",
     "ms": "^2.1.2",
-    "node": "^12.6.0",
+    "node": "^12.10.0",
     "on-finished": "2.3.0",
     "openapi-backend": "^2.3.4",
     "path": "0.12.7",
     "request": "^2.88.0",
-    "rotating-file-stream": "^1.4.2",
-    "rxjs": "^6.5.2",
+    "rotating-file-stream": "^1.4.3",
+    "rxjs": "^6.5.3",
     "serve-favicon": "2.5.0",
-    "simple-get": "3.0.3",
+    "simple-get": "3.1.0",
     "tsscmp": "^1.0.6",
     "url": "0.11.0",
-    "uuid": "3.3.2",
+    "uuid": "3.3.3",
     "winston": "3.2.1"
   },
   "devDependencies": {
     "@openapitools/openapi-generator-cli": "0.0.9-4.0.0-beta3",
     "@types/app-root-path": "^1.2.4",
-    "@types/assert": "^1.4.2",
+    "@types/assert": "^1.4.3",
     "@types/basic-auth": "^1.1.2",
-    "@types/body-parser": "^1.17.0",
-    "@types/chai": "^4.1.7",
-    "@types/compression": "0.0.36",
-    "@types/connect-mongodb-session": "0.0.0",
+    "@types/body-parser": "^1.17.1",
+    "@types/chai": "^4.2.3",
+    "@types/compression": "1.0.1",
+    "@types/connect-mongodb-session": "0.0.1",
     "@types/connect-timeout": "0.0.34",
-    "@types/cookie-parser": "^1.4.1",
-    "@types/debug": "4.1.4",
+    "@types/cookie-parser": "^1.4.2",
+    "@types/debug": "4.1.5",
     "@types/dotenv": "^6.1.1",
     "@types/expect": "^1.20.4",
-    "@types/express": "^4.17.0",
+    "@types/express": "^4.17.1",
     "@types/express-jwt": "0.0.42",
-    "@types/express-session": "^1.15.13",
+    "@types/express-session": "^1.15.14",
     "@types/forever-monitor": "^1.7.4",
     "@types/formidable": "^1.0.31",
-    "@types/http-errors": "^1.6.1",
+    "@types/http-errors": "^1.6.2",
     "@types/js-yaml": "^3.12.1",
     "@types/minimist": "^1.2.0",
     "@types/mocha": "^5.2.7",
-    "@types/mongodb": "^3.1.28",
-    "@types/mongoose": "^5.5.7",
-    "@types/morgan": "^1.7.35",
-    "@types/ms": "^0.7.30",
-    "@types/node": "^12.0.12",
+    "@types/mongodb": "^3.3.2",
+    "@types/mongoose": "^5.5.17",
+    "@types/morgan": "^1.7.37",
+    "@types/ms": "^0.7.31",
+    "@types/node": "^12.7.5",
     "@types/proxyquire": "^1.3.28",
-    "@types/puppeteer": "^1.12.4",
+    "@types/puppeteer": "^1.19.1",
     "@types/puppeteer-core": "^1.9.0",
     "@types/request-promise-native": "^1.0.16",
-    "@types/serve-favicon": "^2.2.30",
+    "@types/serve-favicon": "^2.2.31",
     "@types/shelljs": "^0.8.5",
     "@types/sinon": "^7.0.13",
-    "@types/sinon-chai": "^3.2.2",
+    "@types/sinon-chai": "^3.2.3",
     "@types/uuid": "^3.4.5",
     "@types/winston": "^2.4.4",
-    "acorn": "^6.2.0",
+    "acorn": "^7.0.0",
     "await-repl": "^1.0.2",
     "chai": "^4.2.0",
-    "chrome-launcher": "^0.10.7",
-    "css-loader": "^3.0.0",
-    "del": "^5.0.0",
+    "chrome-launcher": "^0.11.2",
+    "css-loader": "^3.2.0",
+    "del": "^5.1.0",
     "intercept-stdout": "^0.1.2",
     "istanbul": "0.4.5",
     "jsdoc": "^3.6.3",
-    "mocha": "^6.1.4",
+    "mocha": "^6.2.0",
     "mocha-sinon": "2.1.0",
     "nyc": "^14.1.1",
     "prettier": "^1.18.2",
-    "proxyquire": "2.1.0",
-    "puppeteer-core": "^1.18.1",
+    "proxyquire": "2.1.3",
+    "puppeteer-core": "^1.20.0",
     "request-promise-native": "^1.0.7",
     "rewire": "4.0.1",
     "shelljs": "^0.8.3",
-    "sinon": "^7.3.2",
+    "sinon": "^7.4.2",
     "sinon-chai": "3.3.0",
-    "source-map-support": "^0.5.12",
-    "style-loader": "^0.23.1",
+    "source-map-support": "^0.5.13",
+    "style-loader": "^1.0.0",
     "supports-color": "^7.0.0",
-    "ts-loader": "^6.0.4",
-    "ts-node": "^8.3.0",
-    "tslint": "^5.18.0",
+    "ts-loader": "^6.1.0",
+    "ts-node": "^8.4.1",
+    "tslint": "^5.20.0",
     "tslint-eslint-rules": "^5.4.0",
     "tslint-microsoft-contrib": "^6.2.0",
     "tslint-no-unused-expression-chai": "^0.1.4",
-    "typescript": "^3.5.2",
-    "typescript-tslint-plugin": "^0.5.3",
-    "webpack": "^4.35.2",
-    "webpack-cli": "^3.3.5",
+    "typescript": "^3.6.3",
+    "typescript-tslint-plugin": "^0.5.4",
+    "webpack": "^4.40.2",
+    "webpack-cli": "^3.3.9",
     "why-is-node-running": "^2.1.0",
     "wtfnode": "^0.8.0"
   }
